backend/requirements.txt
fastapi
uvicorn
python-multipart
pydub
speechrecognition
google-generativeai
python-dotenv
ffmpeg-python
